refactor: Use single package.json
