+++ .gitignore
@@ -10,7 +10,6 @@
 /electron/coverage
 /electron/dist
 /electron/instrumented-code
-/electron/node_modules
 /electron/renderer/dist
 /keys
 /temp
