--- electron/.npmrc
@@ -1,2 +1,0 @@
-registry=https://registry.npmjs.org/
-save-exact=true
