+++ jenkins/macOS.groovy
@@ -15,7 +15,7 @@ node('master') {
 
   stage('Checkout & Clean') {
     git branch: "${GIT_BRANCH}", url: 'https://github.com/wireapp/wire-desktop.git'
-    sh returnStatus: true, script: 'rm -rf wrap/ electron/node_modules/ node_modules/ *.sig'
+    sh returnStatus: true, script: 'rm -rf wrap/ node_modules/ *.sig'
   }
 
   def text = readFile('info.json')
