+++ package.json
@@ -1,4 +1,5 @@
 {
+  "author": "Wire Swiss <wireapp@wire.com>",
   "build": {
     "directories": {
       "app": "electron",
@@ -7,15 +8,54 @@
     }
   },
   "dependencies": {
+    "@wireapp/certificate-check": "0.0.4",
+    "@wireapp/commons": "1.4.7",
+    "ansi-regex": "4.1.0",
+    "auto-launch": "5.0.5",
+    "axios": "0.18.0",
     "core-js": "2.6.5",
+    "electron-window-state": "5.0.3",
+    "file-url": "2.0.2",
+    "fs-extra": "7.0.1",
+    "image-type": "4.0.0",
+    "jsrsasign": "8.0.12",
+    "lodash": "4.17.11",
+    "logdown": "3.2.8",
+    "minimist": "1.2.0",
+    "moment": "2.24.0",
+    "open-graph": "0.2.4",
+    "prop-types": "15.7.2",
+    "raygun": "0.10.1",
+    "react": "16.8.6",
+    "react-dom": "16.8.6",
+    "react-redux": "6.0.1",
+    "redux": "4.0.1",
+    "redux-logger": "3.0.6",
+    "redux-thunk": "2.3.0",
+    "uuid": "3.3.2",
     "wire-web-config-internal": "https://github.com/wireapp/wire-web-config-default#v0.17.2",
     "wire-web-config-production": "https://github.com/wireapp/wire-web-config-wire#v0.17.3-0"
   },
   "devDependencies": {
     "@babel/core": "7.4.0",
+    "@babel/plugin-proposal-class-properties": "7.4.0",
     "@babel/plugin-proposal-object-rest-spread": "7.4.0",
     "@babel/preset-env": "7.4.2",
     "@babel/preset-react": "7.0.0",
+    "@babel/preset-typescript": "7.3.3",
+    "@babel/register": "7.4.0",
+    "@types/ansi-regex": "4.0.0",
+    "@types/auto-launch": "5.0.1",
+    "@types/file-url": "2.0.0",
+    "@types/fs-extra": "5.0.5",
+    "@types/lodash": "4.14.123",
+    "@types/minimist": "1.2.0",
+    "@types/mocha": "5.2.6",
+    "@types/node": "10.12.29",
+    "@types/open-graph": "0.2.0",
+    "@types/raygun": "0.10.1",
+    "@types/sinon": "7.0.11",
+    "@types/uuid": "3.4.4",
     "@wireapp/copy-config": "0.5.2",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.0.1",
@@ -26,7 +66,9 @@
     "dotenv": "7.0.0",
     "electron": "4.1.3",
     "electron-builder": "20.39.0",
+    "electron-mocha": "6.0.4",
     "electron-packager": "13.1.1",
+    "electron-rebuild": "1.8.4",
     "electron-winstaller": "2.7.0",
     "eslint": "5.16.0",
     "eslint-config-prettier": "4.1.0",
@@ -53,14 +95,16 @@
     "webpack": "4.29.6",
     "webpack-cli": "3.3.0"
   },
-  "license": "LicenseRef-LICENSE",
+  "environment": "internal",
+  "homepage": "https://wire.com",
+  "license": "GPL-3.0",
   "lint-staged": {
-    "*.{js,jsx}": [
-      "eslint --fix",
-      "git add"
-    ],
     "*.ts": [
       "tslint --config tslint.json --project tsconfig.json --fix",
+      "git add"
+    ],
+    "*.{js,jsx}": [
+      "eslint --fix",
       "git add"
     ],
     "*.{json,md,css}": [
@@ -68,48 +112,55 @@
       "git add"
     ]
   },
+  "main": "electron/dist/main.js",
   "name": "wire-desktop",
+  "optionalDependencies": {
+    "@wireapp/node-addressbook": "3.0.2"
+  },
   "private": true,
+  "productName": "WireInternal",
   "repository": {
     "type": "git",
     "url": "https://github.com/wireapp/wire-desktop.git"
   },
   "scripts": {
     "build:linux": "yarn build:ts && grunt linux-prod",
-    "build:linux:internal": "cross-env BUILD_ENV=internal yarn configure && yarn build:ts && grunt linux",
     "build:linux:custom": "yarn build:ts && grunt linux-custom",
+    "build:linux:internal": "cross-env BUILD_ENV=internal yarn configure && yarn build:ts && grunt linux",
     "build:macos": "yarn build:ts && grunt macos-prod",
-    "build:macos:internal": "cross-env BUILD_ENV=internal yarn configure && yarn build:ts && grunt macos",
     "build:macos:custom": "yarn build:ts && grunt macos-custom",
+    "build:macos:internal": "cross-env BUILD_ENV=internal yarn configure && yarn build:ts && grunt macos",
     "build:ts": "yarn clear:ts && tsc",
     "build:win": "yarn build:ts && grunt win-prod",
-    "build:win:internal": "cross-env BUILD_ENV=internal yarn configure && yarn build:ts && grunt win",
     "build:win:custom": "yarn build:ts && grunt win-custom",
-    "bundle:dev": "webpack",
+    "build:win:internal": "cross-env BUILD_ENV=internal yarn configure && yarn build:ts && grunt win",
     "bundle": "webpack --env.production",
+    "bundle:dev": "webpack",
     "clear:ts": "rimraf electron/dist",
     "configure": "copy-config",
+    "coverage": "yarn test",
+    "fix": "yarn fix:js && yarn fix:other && yarn fix:ts",
     "fix:js": "yarn lint:js --fix",
     "fix:other": "yarn prettier --write",
     "fix:ts": "yarn lint:ts --fix",
-    "fix": "yarn fix:js && yarn fix:other && yarn fix:ts",
     "jest": "jest",
+    "lint": "yarn lint:js && yarn lint:other && yarn lint:ts",
     "lint:js": "eslint --ignore-path .gitignore --ext .js,.jsx .",
     "lint:other": "yarn prettier --list-different",
     "lint:ts": "tslint --config tslint.json --project tsconfig.json \"**/*.ts\"",
-    "lint": "yarn lint:js && yarn lint:other && yarn lint:ts",
-    "postinstall": "yarn configure && cd electron && yarn",
+    "postinstall": "electron-rebuild -f -w @wireapp/node-addressbook",
     "prestart": "yarn build:ts && yarn bundle:dev",
     "prettier": "prettier \"**/*.{json,md,css}\"",
+    "start": "cross-env NODE_DEBUG=wire-desktop* electron electron --inspect --devtools --enable-logging",
     "start:dev": "yarn start --env=https://wire-webapp-dev.zinfra.io",
     "start:edge": "yarn start --env=https://wire-webapp-edge.zinfra.io",
     "start:internal": "yarn start --env=https://wire-webapp-staging.wire.com",
     "start:localhost": "yarn start --env=http://localhost:8081",
     "start:prod": "yarn start --env=https://app.wire.com",
     "start:rc": "yarn start --env=https://wire-webapp-rc.zinfra.io",
-    "start": "cross-env NODE_DEBUG=wire-desktop* electron electron --inspect --devtools --enable-logging",
-    "test:electron:main": "cd electron && yarn coverage",
-    "test:react": "jest",
-    "test": "yarn lint && yarn build:ts && yarn test:react && yarn test:electron:main"
-  }
+    "test": "yarn lint && yarn build:ts && yarn test:react && yarn test:electron:main",
+    "test:electron:main": "electron-mocha --require ./babel-register.js \"src/**/*.test?(.main).ts\"",
+    "test:react": "jest"
+  },
+  "updateWinUrl": "https://wire-app.wire.com/win/internal/"
 }
